refactor: Move safe overview tags to its own file
